refactor(Level): general structural improvements
Replaced sf::RectangleShape with sf::Sprite
since they are more suited for textures.

Simplified roomCanBePlaced() to take in a list of sf::IntRect
instead of Room, which then allowed me to remove some member
functions from Room and helper.hpp.

Removed Room::points and Room::rects and
Room now only stores a list of Tiles.

Moved Room::calculateTileTypes to the Room constructor.

The Tiles class was expanded to include a sf::Sprite, which enables
Layer to only hold a single vector of Tiles instead of two separate
for the shape and the type.

Finally, I moved texture loading to the Texture constructor instead
of doing it in Game.